backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic[email]==2.6.0
pydantic-core==2.16.1
pydantic-settings==2.2.1
python-dotenv==1.0.0
openai==1.12.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]>=1.7.5
reportlab==4.0.8
pypdf==3.17.4
httpx==0.26.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.25
alembic==1.13.1
email-validator==2.1.0
beautifulsoup4==4.12.3
aiohttp==3.9.3
firecrawl-py>=4.3.6